Add JSAPI functions for operating on ReadableStream and WritableStream
This adds a few functions for working with ReadableStream, plus introduces a basic set of ones for working with WritableStreams at all.

This patch also includes updates to some of the doc comments for the existing streams API: for many functions the doc comment said that they assert that a passed `Handle<JSObject*>` is (a wrapper for) a specific type, but in reality the function does a check and throws an error instead of asserting. Since this patch introduces a few functions that actually do assert, it seems important to ensure the different behavior is correctly documented.