avoid a weval bug by assigning the iterator to a variable prior to returning
If we return the iterator directly, it somehow ends up with the `kind` parameter being `0` instead of `1`, which changes the iterator behaviour to return the keys instead of the values.

For some reason assigning the iterator to a variable prior to returning makes the `kind` parameter stay the correct value of `1`.