CheckReturn checked the wrong value for isUndefined
To verify that the pbl implementation of CheckReturn matched the version
from InterpreterFrame that it's derived from, I first extracted the
frame->returnValue() call to a variable named retVal to minimize the
differences.

Then it was easier to see that the second `if` was checking
`ret->isUndefined()` instead of `retVal.isUndefined()`, leading to
failure of eight tests in js/src/jit-test/tests/class/.

With this fixed, all but one of the class/ tests pass.