Avoid misaligned jitcode pointer under PBL on wasm.
The PBL tier fills the jitcode pointer on `JSScript`s with an arbitrary
non-null value. This is required because other tiering code uses this
null-check to determine whether it is possible to invoke any higher
tiers (on native platforms, it is set even for baseline interp, to
blinterp's entrypoint). We don't ever actually invoke the code pointer
under PBL, so instead we arbitrarily set it to the function pointer for
PBL's entry point.

The jitcode pointer is stored in the JSScript's cell header, which
requires the low few bits to be clear (they are set when the cell
forwards to another cell after a moving GC pass). On native platforms
function pointers are usually aligned as required, but on wasm32 they
are indices into a table and can have arbitrary values. So instead, we
use a constant value `8` to satisfy the constraint of "arbitrary
non-null value".