pbl: Extend call fastpath to constructor calls
* Make constructor-call fastpath handle primitive returns

EnterJit overrides the frame's return value if it was a constructor call
and the callee returned a primitive instead of an object. Do that in the
return fastpath as well.

* Prevent JitScript from being discarded during CreateThis

* Use rooted reference to callee JSFunction and JSScript

If CreateThis runs, there may be a GC pass, and the callee may move to a
different address. In that case when CreateThis returns we need the GC
to have updated the pointer we're using to the new location, so stash it
in the rooted state.