Undo a change from 89826427d5b34217e1fd55617c14af0cb791ba40, fixing a small perf regression.
The extra conditional doesn't actually seem to be necessary for all
tests to pass even on a debug build. IIRC I had originally added it
because the AfterYield debug hook invocations were not quite right with
respect to an assert, but harmless (one extra call when breakpoints
weren't actually enabled, or something like that); but this doesn't seem
to happen anymore.

For somewhat perplexing reasons, altering this codepath makes a ~6%
performance improvement on native on Octane's Richards benchmark (my
arbitrarily picked test case for a recent regression), even though the
code path never executes. Perhaps it affects other optimizations.