pbl: Make CacheIR input names/types match their definition
No functional changes expected. There are three kinds of changes in this
commit:

- Renaming variables. This should have no effect.

- Changing operand-ID types. These types have the same representation so
  the CacheIR reader interprets the bytecode the same way regardless.

- Re-ordering statements so the CacheIR bytecode is fully consumed
  before the implementation begins.

These changes make it easier to check that each use of the cacheIRReader
matches what's specified in js/src/jit/CacheIROps.yaml. In the future it
may also help with auto-generating this glue code, but that's more
difficult to do correctly so we aren't doing it right now.