pbl: Implement StoreTypedArrayElement (#32)
Many uses of this CacheOp require that GuardToInt32ModUint32 passes
first. Implementing both gives a ~10% speedup for the Mandreel benchmark
from Octane.

The baseline implementation mostly just does a store directly to the
right offset in the array, so calling `SetTypedArrayElement` does extra
work by comparison, but the underlying implementations are all private
to `js/src/vm/TypedArrayObject.cpp` so it's difficult to do better using
public APIs. I got lost trying to figure out why this function needs a
context or GC-rooted handles. So I'm guessing there's a better way to
write this but I'm not sure what it is.