Fix AOT ICs: eagerly ensure JitZone is created.
Somewhere in either the 124..127 rebase or in my move from local JS
shell + wizer hack to StarlingMonkey, the way that Zone initialization
happens has changed and AOT ICs are no longer preloaded. It seems that
the lazy-init creation of the JitZone was only incidentally/fragilely
being triggered during wizening before; this PR adds an explicit
invocation of `ensureJitZone` in the realm creation flow when the AOT
ICs feature is enabled.

Note for future rebases: this should be consolidated into the main AOT
ICs patch.