AOT ICs: rebuild entire corpus.
In the course of the 124..127 upgrade, some IC bodies became invalid
because of altered signatures of CacheIR opcodes. Nominally this should
be fine in a system that never tries to execute such an invalid blob of
bytecode. However, when we pre-weval all bodies in a weval build, we run
into invalid memory reads in the weval phase just as we would if we
tried to execute the IC at runtime in the IC interpreter: fundamentally,
the corpus cannot have invalid bytecode in it. This PR thus removes the
entire corpus and regenerates it:

- Remove `js/src/ics/IC-*`
- Build an engine with `--enable-aot-ics --enable-aot-ics-force
  --enable-aot-ics-enforce`
- Run jit-tests (`./mach jit-test`) and jstests (`./mach jstests`) with
  `AOT_ICS_KEEP_GOING=1`
- Use `js/src/ics/remove-duplicates.py` to remove duplicates among all
  the `IC-*` files in the gecko-dev root (jit-tests) and `js/src/tests`
  (jstests)
- Put all of these files into `js/src/ics`

Note for rebasing in the future: this should be squashed into the main
AOT ICs commit.