PBL: fix specialization range on TableSwitch index.
When instructing weval to generate control flow with
constant-propagated indices for a TableSwitch, we need to give the
proper range; `i - low` has range `0..=(high-low+1)`, not
`low..=high` (i.e. it is shifted down by `low`). Also move the
intrinsic into the not-out-of-bounds conditional to make the code a
little clearer.